% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contexts.R
\name{contexts}
\alias{contexts}
\title{trinucleotide context}
\usage{
contexts(vars, fasta)
}
\arguments{
\item{vars}{mutations with CHROM REF ALT POS columns}

\item{fasta}{genome sequence (from seqinr)}
}
\value{
a vector of contexts for each mutation
}
\description{
trinucleotide context
}
