% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contexts_plot.R
\name{contexts_plot}
\alias{contexts_plot}
\title{plot trinucleotide context}
\usage{
contexts_plot(contexts, full = T)
}
\arguments{
\item{contexts}{results from contexts()}
}
\value{
a list with context_table and a ggplot
}
\description{
plot trinucleotide context
}
