% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_epigenome.R
\name{extract_epigenome}
\alias{extract_epigenome}
\title{Extract values for epigenomic features in a given region. Used in Monroe et al 2022 Not a very generalizable function as it depends on formatting of datasets.}
\usage{
extract_epigenome(
  regions,
  peaks_file,
  meth_file,
  genome_file,
  histone_dir,
  tmp_dir,
  multiBigwigSummary
)
}
\arguments{
\item{regions}{data.table object with columns columns chr, start, stop}

\item{peaks_file}{file with ATACseq peaks}

\item{meth_file}{file with methylation data}

\item{genome_file}{reference genome fasta file}

\item{histone_dir}{directory of files with histone modifications}

\item{tmp_dir}{directory for temporary files}

\item{multiBigwigSummary}{system location for multiBigwigSummary function}
}
\value{
a data.table annotated with genomic and epigenomic features
}
\description{
Extract values for epigenomic features in a given region. Used in Monroe et al 2022 Not a very generalizable function as it depends on formatting of datasets.
}
