% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTajimasD.R
\name{getTajimasD}
\alias{getTajimasD}
\title{extract Tajimas D for features}
\usage{
getTajimasD(vcf, gff)
}
\arguments{
\item{vcf}{directory and file name vcf data}

\item{gff}{directory and file name of reference gff}
}
\value{
writes out a file with tajimnas D values for each feature of a GFF file
}
\description{
extract Tajimas D for features
}
