% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_tss.R
\name{ggplot_tss}
\alias{ggplot_tss}
\title{ggplot_tss}
\usage{
ggplot_tss(tss_means_out, variable)
}
\arguments{
\item{tss_means_out}{results from tss_means}

\item{variable}{which variable are means calulated for}
}
\value{
a vector of contexts for each mutation
}
\description{
ggplot_tss
}
