% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Null_ns_s.R
\name{Null_ns_s}
\alias{Null_ns_s}
\title{get estimate null nonsynonymous/synonymous ratio}
\usage{
Null_ns_s(mutations, composition)
}
\arguments{
\item{mutations}{file with columns 'ref' 'alt' and 'N' counts for each mutation}

\item{composition}{file where each row is coding regions and columns 1,2,3,4, have counts of A, T, C, G}
}
\value{
a list non synonymous/synonymous and stop/synonymous ratios
}
\description{
get estimate null nonsynonymous/synonymous ratio
}
