% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_vcf.R
\name{parse_vcf}
\alias{parse_vcf}
\title{parse through vcf file so it can be used in R}
\usage{
parse_vcf(vcf_file, snpeff = F)
}
\arguments{
\item{vcf_file}{file name vcf data from gatk haplotypecaller}

\item{snpeff}{boolean: the vcf is annotated with snpeff, default = F}
}
\value{
a data.table object with parsed vcf, new columns for various quality meaures, statistics, annotations, etc
}
\description{
parse through vcf file so it can be used in R
}
