% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_counts.R
\name{region_counts}
\alias{region_counts}
\title{add counts of variants to gff data object}
\usage{
region_counts(regions, vcf)
}
\arguments{
\item{regions}{data.table object with columns chr, start, stop}

\item{vcf}{data.table object of variant data (data.table with CHROM, POS, REF, and ALT columns)}
}
\value{
a vector of variant counts for each row of input gff
}
\description{
add counts of variants to gff data object
}
