% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_means.R
\name{tss_means}
\alias{tss_means}
\title{tss_means}
\usage{
tss_means(tss_out = out, variable = "dist", window = 100, na_action = F)
}
\arguments{
\item{tss_out}{results from tss_tts.calc}

\item{variable}{which variable are means calulated for}

\item{window}{window size to summarize in}

\item{na_action}{what to do with NA values in mean()}
}
\value{
a vector of contexts for each mutation
}
\description{
tss_means
}
