% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_tts.feature.R
\name{tss_tts.feature}
\alias{tss_tts.feature}
\title{extract relative position of polymorphisms in relation to gene bodies}
\usage{
tss_tts.feature(
  gff,
  chrs = "all",
  num = "all",
  feature = "gene",
  feature2 = "gene",
  loc = "tss"
)
}
\arguments{
\item{gff}{directory and file name of reference gff OR data.table object of gff data}

\item{chrs}{selection specific chromosomes, default is "all"}

\item{num}{set number of genes, default is "all}

\item{feature}{name of the feature in the gff data that you want to look in relation to, default is "gene"}

\item{feature2}{name of the feature in the gff data that you want test for obverlap default is "gene"}

\item{loc}{"tss" or "tts"}
}
\value{
a vector of the probablity of overlapping in relation to feature
}
\description{
extract relative position of polymorphisms in relation to gene bodies
}
