% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_tts.maf.R
\name{tss_tts.maf}
\alias{tss_tts.maf}
\title{calculate minor allele frequencies of polymorphisms in relation to gene bodies}
\usage{
tss_tts.maf(
  gff,
  vcf,
  type = "both",
  chrs = "all",
  num = "all",
  feature = "gene"
)
}
\arguments{
\item{gff}{directory and file name of reference gff OR data.table object of gff data}

\item{vcf}{directory and file name vcf data OR data.table object of variant data (data.table with CHROM, POS, MAF)}

\item{type}{"snp", "indel", or "both", default is "both"}

\item{chrs}{selection specific chromosomes, default is "all"}

\item{num}{set number of genes, default is "all}

\item{feature}{name of the feature in the gff data that you want to look at, default is "gene"}
}
\value{
a vector of the relative position of polymorphsims in relation to feature
}
\description{
calculate minor allele frequencies of polymorphisms in relation to gene bodies
}
