% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_tts.mappability.R
\name{tss_tts.mappability}
\alias{tss_tts.mappability}
\title{extract relative position of polymorphisms in relation to gene bodies}
\usage{
tss_tts.mappability(gff, mapp, chrs = 1, num = "all")
}
\arguments{
\item{gff}{directory and file name of reference gff OR data.table object of gff data}

\item{mapp}{directory and file name of gen map object}

\item{chrs}{selection specific chromosomes, default is 1}

\item{num}{set number of genes, default is "all}
}
\value{
data.table with pos and D columns
}
\description{
extract relative position of polymorphisms in relation to gene bodies
}
