% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_tts.R
\name{tss_tts}
\alias{tss_tts}
\title{extract relative position of polymorphisms in relation to gene bodies}
\usage{
tss_tts(
  gff = NA,
  vcf = NA,
  loc = "tss",
  type = "snp",
  chrs = "all",
  num = "all",
  feature = "gene"
)
}
\arguments{
\item{gff}{directory and file name of reference gff OR data.table object of gff data}

\item{vcf}{directory and file name vcf data OR data.table object of variant data (data.table with CHROM, POS, REF, and ALT columns)}

\item{loc}{"tss" or "tts", default is "tss"}

\item{type}{"snp", "indel", or "both"}

\item{chrs}{selection specific chromosomes, default is "all"}

\item{num}{set number of genes, default is "all}

\item{feature}{name of the feature in the gff data that you want to look at, default is "gene"}
}
\value{
a vector of the relative position of polymorphsims in relation to feature
}
\description{
extract relative position of polymorphisms in relation to gene bodies
}
