% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_tts.tajima.plot.R
\name{tss_tts.tajima.plot}
\alias{tss_tts.tajima.plot}
\title{plot relative position of tajimas D in relation to gene bodies}
\usage{
tss_tts.tajima.plot(tajima, window = 300, color = "green4")
}
\arguments{
\item{tajima}{object creatd by tss_tts.tajima() with columns pos, D, and loc}

\item{bins}{window size for counting variants, default is 10}
}
\value{
a ggplot object plotting the tss and tts data
}
\description{
plot relative position of tajimas D in relation to gene bodies
}
