% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_tts.tajima.R
\name{tss_tts.tajima}
\alias{tss_tts.tajima}
\title{extract relative position of polymorphisms in relation to gene bodies}
\usage{
tss_tts.tajima(
  gff,
  tajima,
  chrs = "all",
  num = "all",
  loc = "both",
  feature = "gene"
)
}
\arguments{
\item{gff}{directory and file name of reference gff OR data.table object of gff data}

\item{tajima}{directory and file name tajima data OR data.table object of tajima data (data.table with CHROM, BIN_START, N_SNPS, and TajimaD columns)}

\item{chrs}{selection specific chromosomes, default is "all"}

\item{num}{set number of genes, default is "all}

\item{loc}{"tss" or "tts"}

\item{feature}{name of the feature in the gff data that you want to look in relation to, default is "gene"}
}
\value{
data.table with pos and D columns
}
\description{
extract relative position of polymorphisms in relation to gene bodies
}
