% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_tts.variants.plot.R
\name{tss_tts.variants.plot}
\alias{tss_tts.variants.plot}
\title{plot relative position of polymorphisms in relation to gene bodies}
\usage{
tss_tts.variants.plot(positions, window = 10, color = "green4")
}
\arguments{
\item{positions}{object creatd by tss_tts.variants() with columns pos and loc}

\item{window}{window size for counting variants, default is 10}
}
\value{
a ggplot object plotting the tss and tts data
}
\description{
plot relative position of polymorphisms in relation to gene bodies
}
